Fix goroutine leaks.
The wait function was leaking a goroutine via time.Tick.

Also—and I'm not sure this was an actual problem—Cmd.Start appers to create goroutines to be cleaned up by Cmd.Wait. Since we never called Cmd.Wait, these will be leaked too. So, we just call cmd.Wait as a dummy after Wait4.

Fixes #20.